+++ packages/@headlessui-vue/src/components/listbox/listbox.ts
@@ -161,8 +161,17 @@ export let Listbox = defineComponent({
         searchQuery.value = ''
       },
       registerOption(id: string, dataRef: ListboxOptionDataRef) {
+        let orderMap = Array.from(
+          optionsRef.value?.querySelectorAll('[id^="headlessui-listbox-option-"]') ?? []
+        ).reduce(
+          (lookup, element, index) => Object.assign(lookup, { [element.id]: index }),
+          {}
+        ) as Record<string, number>
+
         // @ts-expect-error The expected type comes from property 'dataRef' which is declared here on type '{ id: string; dataRef: { textValue: string; disabled: boolean; }; }'
-        options.value.push({ id, dataRef })
+        options.value = [...options.value, { id, dataRef }].sort(
+          (a, z) => orderMap[a.id] - orderMap[z.id]
+        )
       },
       unregisterOption(id: string) {
         let nextOptions = options.value.slice()